% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic-events.R
\name{brew_analytics_events}
\alias{brew_analytics_events}
\alias{brew_install_events}
\alias{brew_install_on_request_events}
\alias{brew_build_error_events}
\alias{brew_cask_install_events}
\alias{brew_os_version_events}
\title{Retrieve Analytics Events for Homebrew Formulae for a Given Category}
\usage{
brew_analytics_events(category = c("install", "install-on-request",
  "cask-install", "build-error", "os-version"), delta = c(30, 90, 365))

brew_install_events(delta = c(30, 90, 365))

brew_install_on_request_events(delta = c(30, 90, 365))

brew_build_error_events(delta = c(30, 90, 365))

brew_cask_install_events(delta = c(30, 90, 365))

brew_os_version_events(delta = c(30, 90, 365))
}
\arguments{
\item{category}{the category of the analytics events (the default is \code{install} events.
The following are acceptable inputs:
\itemize{
\item \code{install}: the installation of all formulae
\item \code{install-on-request}: the requested installation of all formulae (i.e. not as a dependency of other formulae)
\item \code{cask-install}: the installation of all casks
\item \code{build-error}: the installation failure of all formulae
\item \code{os-version}: the macOS version of the machine that submitted an event
}}

\item{delta}{number of previous days to return. One of \code{30}, \code{90}, \code{365}}
}
\value{
data frame (tibble)
}
\description{
Get analytics event data such as event count, percent and time range.
}
\details{
The old Homebrew Analytics API had different endpoints for analytics
event metrics. These have been kept as wrappers to the new function and
other conveneince functions have been added. Thse are:
\itemize{
\item \code{[brew_build_error_events()]}
\item \code{[brew_cask_install_events()]}
\item \code{[brew_install_events()]}
\item \code{[brew_install_on_request_events()]}
\item \code{[brew_os_version_events()]}
}
}
\note{
The results of the API call are memoised as they won't likely
won't change between calls in the same R session. Use the
facilities provided by the \code{memoise} package to clear the
cache if you use this in a long running R session.
}
\examples{
xdf <- brew_analytics_events("install", 30)
}
\references{
\url{https://formulae.brew.sh/docs/api/}
}
