% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-error.R
\name{brew_build_error_events}
\alias{brew_build_error_events}
\title{Retrieve Build Errors for the past 30 days for Homebrew Formulae}
\usage{
brew_build_error_events()
}
\value{
data frame (tibble) with build recipe and count;
also includes a \code{meta} attribute with the metrics category and
metrics tart/end date.
}
\description{
Retrieve Build Errors for the past 30 days for Homebrew Formulae
}
\examples{
xdf <- brew_build_error_events()
attr(xdf, "meta")
}
\references{
\url{https://formulae.brew.sh/analytics/}
}
