% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew-formulae.R
\name{brew_formulae}
\alias{brew_formulae}
\title{Retrieve formulae metadata for all homebrew-core formulae}
\usage{
brew_formulae(formulae = "all")
}
\arguments{
\item{formluae}{if "\code{all}" (the default) then metadata on all
Homebrew/homebrew-core formulae will be returned. Otherwise
the API will be queried for all formulae in the passed in
character vector and returned, skipping (and warning) on
any specified formula that does not exist.}
}
\value{
data frame (tibble) with formula analytics data or \code{NULL} if
what you looked for was not found.
}
\description{
Get the `brew info --json=v1`` output for \emph{all} current
Homebrew/homebrew-core formulae \emph{or} get the same output for
one or more current Homebrew/homebrew-core formulae with an extra
analytics key with analytics data by specify
}
\note{
The results of the API call are memoised as they won't likely
won't change between calls in the same R session. Use the
facilities provided by the \code{memoise} package to clear the
cache if you use this in a long running R session.
}
\examples{
f <- brew_formulae() # same as manually specifying "all"
f <- brew_formulae("curl")
f <- brew_formulae(c("curl", "youtube-dl", "blemenge"))
}
\references{
\url{https://formulae.brew.sh/docs/api/}
}
