% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-on-request-events.R
\name{brew_install_on_request_events}
\alias{brew_install_on_request_events}
\title{Retrieve Install-on-Requests nts for Homebrew Formulae}
\usage{
brew_install_on_request_events(delta = c(30, 90, 365))
}
\arguments{
\item{delta}{number of previous days to return. One of \code{30}, \code{90}, \code{365}}
}
\value{
data frame (tibble) with formula, configure options and count;
also includes a \code{meta} attribute with the metrics category and
metrics tart/end date.
}
\description{
Retrieve Install-on-Requests nts for Homebrew Formulae
}
\examples{
xdf <- brew_install_on_request_events(30)
attr(xdf, "meta")
}
\references{
\url{https://formulae.brew.sh/analytics/}
}
