% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homebrewanalytics-package.R
\docType{package}
\name{homebrewanalytics}
\alias{homebrewanalytics}
\alias{homebrewanalytics-package}
\title{homebrewanalytics}
\description{
The 'Homebrew Project' <brew.sh> has a myriad of "recipes" that
make life easier for 'macOS' users by enabling them to (mostly) effortlessly
install popular open source libraries, tools, utilities and applications.
The project collectes anonymous metrics from users who have not opted-out
of metrics collection and makes them available via a 'JSON' 'API'.
}
\author{
Bob Rudis (bob@rud.is)
}
